Consolidate Harvey UFS headers into ufsdat.h and ufsfns.h
Still need to tidy a bit more -
 - should consolidate harvey UFS .c files
 - should tidy contents of new headers
 - need to look at libufs again

Once that's done, I'll get onto reading a file...
