+++ sys/include/ufs/inode.h
@@ -177,5 +177,3 @@ struct ufid {
 	uint32_t  ufid_ino;	/* File number (ino). */
 	uint32_t  ufid_gen;	/* Generation number. */
 };
-
-Vtype ifmt_to_vtype(uint16_t imode);
