+++ sys/src/9/ufs/ffs_alloc.c
@@ -65,9 +65,9 @@
 #include "dat.h"
 #include "fns.h"
 
+#include "ufs/ufsdat.h"
 #include <ufs/libufsdat.h>
 #include <ufs/freebsd_util.h>
-#include "ufs_harvey.h"
 
 //#include <ufs/ufs/dir.h>
 //#include <ufs/ufs/extattr.h>
