+++ sys/src/9/ufs/ffs_snapshot.c
@@ -39,9 +39,10 @@
 #include "dat.h"
 #include "fns.h"
 
+#include "ufsdat.h"
 #include <ufs/libufsdat.h>
+#include "ufsfns.h"
 #include <ufs/freebsd_util.h>
-#include "ufs_harvey.h"
 
 //#include <ufs/ufs/extattr.h>
 #include "ufs/quota.h"
