+++ sys/src/9/ufs/ffs_softdep.c
@@ -45,9 +45,9 @@
 #include "dat.h"
 #include "port/portfns.h"
 
+#include "ufs/ufsdat.h"
 #include <ufs/libufsdat.h>
 #include <ufs/freebsd_util.h>
-#include "ufs_harvey.h"
 
 /*
  * For now we want the safety net that the DEBUG flag provides.
