+++ sys/src/9/ufs/ffs_vnops.c
@@ -69,7 +69,6 @@
 
 #include <ufs/libufsdat.h>
 #include <ufs/freebsd_util.h>
-#include "ufs_harvey.h"
 
 //#include <ufs/ufs/extattr.h>
 //#include <ufs/ufs/quota.h>
