+++ sys/src/9/ufs/ufs_harvey.c
@@ -15,11 +15,10 @@
 #include "dat.h"
 #include "fns.h"
 
+#include "ufsdat.h"
 #include <ufs/libufsdat.h>
+#include "ufsfns.h"
 #include <ufs/freebsd_util.h>
-#include "ufs_harvey.h"
-#include "ufs_mountpoint.h"
-#include "ufs_ext.h"
 
 #include "ufs/quota.h"
 #include "ufs/inode.h"
