+++ sys/src/9/ufs/ufs_vnops.c
@@ -40,9 +40,10 @@
 #include "dat.h"
 #include "port/portfns.h"
 
+#include "ufsdat.h"
 #include <ufs/libufsdat.h>
+#include "ufsfns.h"
 #include <ufs/freebsd_util.h>
-#include "ufs_harvey.h"
 
 #include "ufs/quota.h"
 #include "ufs/inode.h"
